Com visao detalhada.

+++ lib/templates/rstatic/rstatic_template.rb
@@ -1,5 +1,6 @@
 require 'fileutils'
 require 'syntax/convertors/html'
+require 'ruby_static'
 
 class RstaticTemplate < MetricFu::Template
 
@@ -24,31 +25,34 @@ class RstaticTemplate < MetricFu::Template
     convertor = Syntax::Convertors::HTML.for_syntax('ruby')
 
     per_file_data.each_pair do |file, lines|
-      data = File.open(file, 'r').readlines
-      fn = "#{file.gsub(%r{/}, '_')}.html"
+      if File.exist?(file)
+        data = File.open(file, 'r').readlines
+        fn = "#{file.gsub(%r{/}, '_')}.html"
 
-      out = "<html><body>"
-      out << "<table cellpadding='0' cellspacing='0' class='ruby'>"
-      data.each_with_index do |line, idx|
-        out << "<tr><td valign='top'><small>#{idx + 1}</small></td>"
-        out << "<td valign='top'>"
-        if lines.has_key?((idx + 1).to_s)
-          out << "<ul>"
-          lines[(idx + 1).to_s].each do |problem|
-            out << "<li>#{problem[:description]} &raquo; #{problem[:type]}</li>"
+        out = "<h3>#{fn}</h3></br>"
+        out << "<div class='code_table'>"
+        out << "<table cellpadding='0' cellspacing='0' class='ruby'>"
+        data.each_with_index do |line, idx|
+          out << "<tr><td valign='top' class='line_number'><small>#{idx + 1}</small></td>"
+          out << "<td valign='top' class='score'>"
+          if lines.has_key?((idx + 1).to_s)
+            out << "<ul>"
+            lines[(idx + 1).to_s].each do |problem|
+              out << "<li>#{problem[:description]} &raquo; #{RubyComplexity::NAME}</li>"
+            end
+            out << "</ul>"
+          else
+            out << "&nbsp;"
           end
-          out << "</ul>"
-        else
-          out << "&nbsp;"
+          out << "</td>"
+          line_for_display = MetricFu.configuration.syntax_highlighting ? convertor.convert(line) : line
+          out << "<td valign='top' class='code'><a name='line#{idx + 1}'>#{line_for_display}</a></td>"
+          out << "</tr>"
         end
-        out << "</td>"
-        line_for_display = MetricFu.configuration.syntax_highlighting ? convertor.convert(line) : line
-        out << "<td valign='top'><a name='line#{idx + 1}'>#{line_for_display}</a></td>"
-        out << "</tr>"
-      end
-      out << "<table></body></html>"
+        out << "</table></div>"
 
-      MetricFu.report.save_output(out, MetricFu.output_directory, fn)
+        MetricFu.report.save_output(out, MetricFu.output_directory, fn)
+      end
     end
   end
